Spirit: merging fixes for warnings and missing header #include's from trunk
[SVN r57475]

+++ test/lex/set_token_value.cpp
@@ -121,7 +121,7 @@ bool test_tokens(token_data const* d, std::vector<Token> const& tokens)
             return false;           // reached end of expected data
 
         typename Token::token_value_type const& value (t.value());
-        if (t.id() != d->id)        // token id must match
+        if (t.id() != static_cast<std::size_t>(d->id))        // token id must match
             return false;
         if (value.which() != 1)     // must have an integer value 
             return false;
@@ -140,7 +140,7 @@ bool test_indents(int *i, std::stack<unsigned int>& indents)
     {
         if (*i == -1)
             return false;           // reached end of expected data
-        if (indents.top() != *i)
+        if (indents.top() != static_cast<unsigned int>(*i))
             return false;           // value must match
 
         ++i;
